requirements.txt
keras==2.15.0
matplotlib==3.7.1
numpy==1.26.4
opencv_python==4.9.0.80
opencv_python_headless==4.8.1.78
Pillow==9.5.0
Pillow==10.3.0
tensorflow==2.15.0
pydot==2.0.0
